Adding command directive to authorized_keys entry
This will add the respective command directive to the
authorized_keys entry, which is to be added on the keyserver. This
way the key cannot be used to spawn a shell or execute any other
command. The key can only be used to fetch the LUKS key from the
server.